% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockMatrixDiagonal.R
\name{blockMatrixDiagonal}
\alias{blockMatrixDiagonal}
\title{diagonal block matrix}
\usage{
blockMatrixDiagonal(...)
}
\arguments{
\item{...}{a list of matrices}
}
\value{
diagonal block matrix concatinated from this list of matrices
}
\description{
create a diagonal block matrix
}
\examples{
m1<-matrix(round(runif(4*4),1),nrow=4,ncol=4)
m2<-matrix(round(runif(4*4),1),nrow=4,ncol=4)
blockMatrixDiagonal(m1,m2,m2,m1)

sigma.1<-0.1
sigma.2<-0.4
J<-10 #subjects
I<-3 #cluster
V.i<-sigma.2*matrix(1,nrow=J,ncol=J)+sigma.1*diag(1, nrow=J,ncol=J) #Covarianmatrix of one cluster
x<-lapply(1:I, function(X) V.i)
blockMatrixDiagonal(x) #Covarianmatrix of all cluster
}

