% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designMatrix.R
\name{completeDataDesignMatrix}
\alias{completeDataDesignMatrix}
\title{Design matrix for complete data within design}
\usage{
completeDataDesignMatrix(J, X)
}
\arguments{
\item{J}{number of subjects}

\item{X}{given design matrix}
}
\value{
design matrix for complete data within design
}
\description{
create design matrix for complete data within design
}
\examples{
K<-6  #measurement (or timepoints)
I<-10 #Cluster
J<-2 #number of subjects
X<-designMatrix(nC=I, nT=K, nSw=2)
completeDataDesignMatrix(J, X)
}

