% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovarianceMatrix.R
\name{CovMat_Design}
\alias{CovMat_Design}
\title{covariance matrix for the multivriate normal distributed variables}
\usage{
CovMat_Design(K, J, I, sigma.1.q, sigma.2.q = NULL, sigma.3.q)
}
\arguments{
\item{K}{number of timepoints or measurments (design parameter)}

\item{J}{number of subjects}

\item{I}{number of clusters (design parameter)}

\item{sigma.1.q}{variance of the lowest level (error variance or within subject variance)}

\item{sigma.2.q}{secound level variance (e.g. within cluster and between subject variance)}

\item{sigma.3.q}{third level variance (e.g. between cluster variance)}
}
\value{
covariance matrix
}
\description{
covariance matrix fof the normal distribution under cluster randomized study type given a design and a type
}
\examples{
K<-6  #measurement (or timepoints)
I<-10 #Cluster
J<-2 #number of subjects

sigma.1<-0.1
sigma.3<-0.9
CovMat_Design(K, J, I,sigma.1.q=sigma.1, sigma.3.q=sigma.3)

sigma.1<-0.1
sigma.2<-0.4
sigma.3<-0.9
CovMat_Design(K, J, I,sigma.1.q=sigma.1, sigma.2.q=sigma.2, sigma.3.q=sigma.3)
}

