% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designMatrix.R
\name{designMatrix.SWD}
\alias{designMatrix.SWD}
\title{Design matrix for SWD model}
\usage{
designMatrix.SWD(nC, nT, nSw)
}
\arguments{
\item{nC}{number of cluster}

\item{nT}{number of timepoints}

\item{nSw}{number of cluster switches per time point from control to intervention}
}
\value{
design matrix for a given setup of a stepped wedge design
}
\description{
create design matrix for a given setup of a stepped wedge design
}
\examples{

designMatrix.SWD(5,6,1)

K<-6  #measurement (or timepoints)
I<-10 #Cluster
designMatrix.SWD(nC=I, nT=K, nSw=2)

}

