% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovarianceMatrix.R
\name{CovMat_Design}
\alias{CovMat_Design}
\title{covariance matrix for the normal distribution}
\usage{
CovMat_Design(K, J, I, sigma.1, sigma.2 = NULL, sigma.3, design, type)
}
\arguments{
\item{K}{number of timepoints or measurments (design parameter)}

\item{J}{number of subjects}

\item{I}{number of clusters (design parameter)}

\item{sigma.1}{variance of the lowest level (error variance or within subject variance)}

\item{sigma.2}{secound level variance (e.g. within cluster and between subject variance)}

\item{sigma.3}{third level variance (e.g. between cluster variance)}

\item{design}{is study design, possible "SWD", "case-control" or "cross-over"}

\item{type}{is either cross-sectional or longitudinal}
}
\value{
covariance matrix
}
\description{
covariance matrix fof the normal distribution under cluster randomized study type given a design and a type
}
\examples{
K<-6  #measurement (or timepoints)
I<-10 #Cluster
J<-2 #number of subjects

sigma.1<-0.1
sigma.3<-0.9
CovMat_Design(K, J, I,sigma.1=sigma.1, sigma.3=sigma.3, design="SWD", type="cross-sec")

sigma.1<-0.1
sigma.2<-0.4
sigma.3<-0.9
CovMat_Design(K, J, I,sigma.1=sigma.1, sigma.2=sigma.2, sigma.3=sigma.3, design="SWD", type="long")
}

